chore: refactor home page layout to remove unnecessary Expanded widget
The Expanded widget was removed from the home page layout as it was causing layout issues. The Column widget now directly contains the Text widget and the ButtonGrid widget is commented out.

feat: add support for customizing button size in ButtonGrid widget

The ButtonGrid widget now uses the ButtonTheme widget to customize the size of the ElevatedButton widgets. The minWidth and height properties of the ButtonTheme widget are set to 100.0 and 50.0 respectively.

fix: import missing dependencies in profile page

The profile page was missing imports for the MyOutlineButton widget and the RegisterPage widget. These imports have been added to the profile page.

feat: add "Daftar Akun" and "Masuk Akun" buttons to profile page

Two MyOutlineButton widgets have been added to the profile page. The first button is labeled "Daftar Akun" and when pressed, it navigates to the RegisterPage. The second button is labeled "Masuk Akun" and when pressed, it navigates to the LoginPage.

chore: update text and style in search page

The text and style in the search page have been updated. The text now says "Search Page is Under Construction !!" and the style of the text is set to bold.

feat: add heading and description to tourist spot detail page

The tourist spot detail page now displays the name and description of the tourist spot. The name is displayed using the AppStyles.headingStyle and the description is displayed below it.

chore: update font size in AppStyles

The AppStyles class now includes a constant for xtraLargeFontSize with a value of 32.0. This font size can be used for headings.